Large update of random number generator implementations
Summary:
While working on a bug in the CEEMD attribute, many
issues were raised in the current implementations:
1- A different implementation is used on Unix and Windows, both being
described as obsolete in the documentation of the function currently being
used. Hence the choice to replace them with function introduced in C++11,
the same on all platforms
2- Made the user-defined seed a class member of each generator, not a
global parameter. However I tried to make it clear in the documentation
that the seed should not be set in production code, as a much better seed
is set when the generator gets constructed.
3- For the uniform number generator, two distinct generators are set,
one for drawing 4 bytes integers, a second for drawing long int and double types,
to fully use their 8 bytes dynamic range. This latter is the one recommended
by the standard C++ documentation.
4- The previous implementation of RandGen::getInt() was never drawing negative
integers. For me this is a bug, RandGen::getIndex should be used if one wants
to draw a strictly positive integer. Could not find any use of it in our code.
5- The functions RandGen::getIndexFast have severe shortcomings: have drawing
about 120 values with their current implementations, all subsequent drawn values
are equal to 1. Furthermore when looking at their actual runtime they perform
slower than their non-fast versions. Hence I propose to deprecate them.
6- The global function Stats::randGen() did not return the reference to the object

[Jira: OD-3859]

Test Plan:
Made a thourough test program to highlight the differences in
quality and implementation between the implementation in main and 6.6 branch.
The test program is actually is very good combined test of not only the
random number generators, but also the RunCalc, ParallelCalc and core functions
inside the ArrayMath namespace.
The precision with the RunCalc and ArrayMath functions is about the same between
branches, the precision of the ParallelCalc is clearly demonstrated as inferior for
now, because that class cannot process cumulative sums in a precision larger than
the input array precision (double for float array, long double for double array)
Runtime slighty higher but acceptable with the uniform generator,
runtime is almost identical with the normal generator.

Reviewers: raman

Reviewed By: raman

Subscribers: nanne

Differential Revision: https://phabricator.dgbes.com:8443/D14819