Fixed crash when restoring many pseudowells in SynthRock
Summary:
The StratPropSyntheticDataCreator was recently made
multi-threaded, and uses a FFT filter in its doWork implementation.
At the same time a HiddenParameter was added to ArrayNDWindow, to support
different taper parameter value per direction, as needed in the Seis::DataGlueer
used by the Machine Learning plugin.
The FFTFilter object was constructed inside the for loop of StratPropSyntheticDataCreator::doWork,
leading in multiple instances of HiddenParam<ArrayNDWindow,TypeSet<float>*> hp_paramval_
being accessed simultaneously.
That creates a repeatible crash, but only if a very large number of pseudowells
are created/restored, on a machine with many CPUs (dgb33 used 16 CPUs).
To mitigate this risk on this branch, and although it is not common practise,
I have disabled the hidden parameter when the ArrayNDWindow is 1D,
and moved the construction of the FFTFilter outside of the for loop of
StratPropSyntheticDataCreator::doWork, eventually into
StratPropSyntheticDataCreator::doPrepare where the objects are guaranteed to be
created sequentially.
In the main branch it can remain in the doWork function, but not in its for loop.

[Jira: OD-3873]

Test Plan:
Restoring Hardeep's 70000 pseudowells now works without crash,
tested applying a FFT Filter attribute on-the-fly (1D CosTaper), and
tested running a Fault volume prediction from the ML plugin (3D CosTaper with
variable parameters).
No crash, correct results.

Reviewers: nanne

Reviewed By: nanne

Subscribers: raman

Differential Revision: https://phabricator.dgbes.com:8443/D14831