Many improvements for distributed processing [devel]
Summary:
To start with, added public functions to JobIOHandler, JobIOMgr,
and JobRunner to export the authority of the server that start listening on
the network for connections from remote nodes. It is this authority, split
into port and host, that must be used on the command line when submitting work
to a remote machine.
With that reachable from uiMMBatchJobDispatcher, added a QC in
uiMMBatchJobDispatcher::startWork and debug log output to have it state
which authority the server is listening too, to ensure this is no longer
silently listens on a loopback interface type of address.

Pushed the current configuration DNS vs. IP address one step further:
The connection is now done either by IP or by host name depending on the
node configuration. Made it such that the GUI will display a clear HostData
string that indicates which connection type is used, by display first the
primary name (DNS or IP), and thereafter the counterpart.

Fixed a crash when closing the Distributed processing window after the merge
was already done, as it tried to do it again (but the temporary storage was gone).
Added the required event loop to the od_remexec app, required to resolve host names.

[Jira: OD-2958]

Test Plan:
Tested single machine processing to local host or remote Linux host,
and distributed processing to both windows and Linux, using various configurations:
1- DNS based with/without the local host
2- IP based with/without the local host
3- Mix of DNS and IP based nodes.
Checked by setting DTECT_DEBUG to 0x0020 that the issued commands
had the right format (IP or host name) for both the client machine argument and
the remote node argument.
Could only do that from a OpenSuse Linux machine, will need a release to test it
from either Windows or CentOS7.

Reviewers: nanne

Reviewed By: nanne

Subscribers: raman

Differential Revision: https://phabricator.dgbes.com:8443/D14836