Deprecated the global random number generator [devel]
Summary:
As mentioned before, should be replaced by local
or class member objects. The second should be preffered if it is
likely that the class will use the generator for its standard work.

Test Plan:
clippping on displayed inlines seems OK, histogram computation
in the seismic pre-loading as wells.
Everything compiles without warning.

Reviewers: nanne

Reviewed By: nanne

Subscribers: raman

Differential Revision: https://phabricator.dgbes.com:8443/D14844