Fixed distributed processing not starting from Windows [devel]
Summary:
Now that the authority on which the server is listening is
given to uiMMBatchJobDispatcher, the newly added checks discard it on
windows because the corresponding hostname is equal to the 'localhost' string.
I traced back the origin of this wrong host name inside the authority class to
this commit:
b5cf227

All this added code is wrong (thus reverted), and is actually not the core
issue, which is coming from the enum comparison in Network::Server::authority(),
when comparing the protocol with an enum value. Although
 QAbstractSocket::UnknownNetworkLayerProtocol is set to -1, the comparison returns false on windows (but true on Linux), unless both compared values are strictly casted to integers.

To ensure that we can reliably convert QHostAddress::SpecialAddress enums to
authorities and their hostnames, I added a test function to the systeminfo class
It must check constructing the authority object from these special addresses (used when a listening server is started), but also IP addresses and hostnames,
either loopback interfaces or IP type of interfaces/addresses.

[Jira: OD-2958]

Test Plan:
test program runs on windows where this code is written,
and distributed processing starts. Needs to be tested after commit on Linux.