Implement UI to manage od_remoteservice
Summary:
Fixes OD-3877 Rewrite the od_remote_service_manager code
I've incorporated the proposed changes in revision D13179 and Nanne's suggestion
to introduce a uiTrayDialog class.
The current implementation of the od_remoteservicemgr UI doesn't implement any IPC with
the od_remoteservice daemon. It's possible that using signals might be sufficient if
handled properly from the daemon side and all that needs to be done is to kill/stop it.

Test Plan:
Tested on Linux:
- only observed starting and stopping of the daemon in the task list - didn't try to set
up an MMProc system

Reviewers: nanne, arnaud

Reviewed By: arnaud

Differential Revision: https://phabricator.dgbes.com:8443/D14857