Showing standard sort indicator in XPlot table
Summary:
Added uiTable::setColumnSortIndicator.
If sort indicator is enabled, by default it is shown on the current column. I had to override that by calling the function each time a column is selected.

Test Plan: Sort indicators are shown correctly on the currently sorted column. Selecting the column as X/Y axis does not affect the sort indicator anymore.

Reviewers: nanne

Reviewed By: nanne

Differential Revision: https://phabricator.dgbes.com:8443/D14875