Fixed Distributed processing for some DNS configurations [devel]
Summary:
Between 6.6.7rc2 and 6.6.7rc3 I started making using of the
Network::Authority from the listening server of JobIOMgr when setting up
the machine command to be executed on the remote node. One of the argument
values in this machine command is the address to be used by the nodes for
connecting back to od_SeisMMBatch. Previously it was always an IPv4 address,
and I switched it to a host name. See:
https://github.com/OpendTect/OpendTect/commit/320a6d6233b0135cc89ea90318c018a10245f88d
(line 697 of src/MMProc/jobiomgr.cc)

The value is taken from Network::Authority::getHost() which in the case of
a listening server was equivalent to GetLocalHostName(). I propose the following:
1- The function System::localFullHostName() is added to partly replace the
usage of GetLocalHostName() and System::localHostName()
It should return a fully qualified host name, providing that such domain is setup
(there is not always one)
2- Its implementation should read the hostname overrule from the user settings,
and use it (with some QC to check that it effectively matches the local machine)
3- All occurences of GetLocalHostName() which will be used for making a
connection should be switched to System::localFullHostName(), and also the host name
stored in the HostData class when setup in DNS mode.
4- Logging, and directories annotated by host name string should be using
System::localHostNameWoDomain(), cleaned up from domain name, because
localHostName() may return a name that contains the domain name on RedHat/CentOS

[Jira: OD-3909]

Test Plan:
Could already check processing between two windows machines
on a network without domain name, between the patched od_SeisMMBatch and
a node running the daemon on either the public 6.6.6 or 6.6.7rc5 releases
More tests to be done at the office.

Reviewers: raman, nanne

Reviewed By: nanne

Subscribers: wayne

Differential Revision: https://phabricator.dgbes.com:8443/D14882