Fix OD3800 crash in uiSVF (Thalweg Tracker) when closing OpendTect
Summary:
Fixes OD-3800
This is a very repeatable crash (on Linux anyway, user reported it on Windows as well) that occurs
when shutting down OpendTect with the uiSVF (Thalweg Tracker) dialog open. I thought I could fix it
by replacing the old style callback-notify class with mAttachCB and detachAllNotifier calls in uiSVF
and uiVisPartServer but it didn't work. The crash occurs because the objectRemoved notifier in
uiVisPartServer has a non-empty cbs_ (CallBackSet) with a negative reference count.

Any ideas?

Test Plan: Tested on Linux

Reviewers: raman, nanne

Reviewed By: nanne

Subscribers: nanne

Differential Revision: https://phabricator.dgbes.com:8443/D14879