Add setting to control uiTable sorting [devel]
Summary:
Noticed some funny behaviour in the batch hosts dialog. Clicking in the table header
was sorting the table contents but the view wasn't changing. Selecting the row delete icon
was showing a confirmation message based on the sorted contents that was at odds with what
the user was seeing. Added the ability to turn off uiTable sorting.

[Jira: OD-3933]

Test Plan: Tested on Linux

Reviewers: arnaud

Reviewed By: arnaud

Subscribers: nanne

Differential Revision: https://phabricator.dgbes.com:8443/D14923