Fix for host selection in BatchHosts dialog
Summary:
Jira: OD-3933
I made a few more improvements in this dialog, but the real fix is in
uiTable where the focus of the uiComboBox should be disabled.

Enabled multi host remove and fixed button sensitivity in readonly mode.

Test Plan: Selection works properly now and message shows the names of the selected hosts.

Reviewers: arnaud

Reviewed By: arnaud

Subscribers: wayne

Differential Revision: https://phabricator.dgbes.com:8443/D14926