Use TCP to control MMProc daemons
Summary:
Fixes OD-3923
I tried to use the existing ServiceClientMgr and Service ServerMgr classes to implement
the IPC between the MMProc daemons and od_BatchHosts and od_RemoteServiceMgr but couldn't
get it working for exchange of messages with "rich content" (ie more than Ok or Err). So
I created 2 standalone classes, MMPServer and MMPServerClient, and have used them to provide
IPC and control of the daemons via JSON packets. Current status:
- od_RemoteServiceMgr now takes over a od_remoteservice daemon should it be already running
- od_RemoteServiceMgr can change the DataRoot set for the od_remoteservice daemon
- od_BatchHosts all the previous direct socket calls have been replaced by the new MMPServerClient
class
- od_remoteservice now uses a built in MMPServer. IPC hooks exist to get/set/check the DataRoot get
the daemon status and start jobs
- od_remexec now uses MMPServerClient to communicate with a remote od_remote_service daemon. At this
stage only a StartJob command is implemented.

Only limited testing has been undertaken at this point. Plan is to pull down the patch for this revision
and build it on Windows for proper multi-machine testing.

Test Plan: Limited testing on Linux:

Reviewers: arnaud

Reviewed By: arnaud

Subscribers: raman

Differential Revision: https://phabricator.dgbes.com:8443/D14936