Fixed crash in SynthRock with async zoom mode
Summary:
the multi flatviewer control should not delete the flatviewer
annotations in its destructor, because the flatviewer desctructor will
take care of it. For safety, added a callback that sets the pointers
in zoomboxes_ to null when the flatviewer is deleted.

Test Plan: no more crash using the workflow in OD-3971

Reviewers: nanne

Reviewed By: nanne

Subscribers: raman

Differential Revision: https://phabricator.dgbes.com:8443/D14956