Some fixes for od6.6 CRS test ptogram [devel]
Summary: Some tests were not running because when an array is assigned to a pointer, the sizeof(ptr) gives the size of the Pointer and not the size of the original array. Incidentally these tests would fail as the values were slightly off. So, chnaged the values as well. Also made the test description consistent.

Test Plan: All the tests are running successfully.

Reviewers: arnaud

Reviewed By: arnaud

Differential Revision: https://phabricator.dgbes.com:8443/D14965