Fixed auto-loading of commercial plugins
Summary:
The auto-loading of plugins can be conditioned by the filter of
the plugin selection window when starting OpendTect. However since
this class was writing the selection to the user settings, the selection
would apply to all OpendTect apps and not only od_main, and that is
especially harmful for batch programs that will have missing runtime libraries.

Added a flag to PluginManager::loadAuto, to ensure that the stored list is
exclusively used when a uiPluginSel object is used. In all other cases, which
means also when od_main is started but the plugin selection window is disabled
by the environment variable OD_SKIP_PLUGIN_SEL, all required plugins will be
loaded when starting any OpendTect application.

[Jira: OD-3921]

Test Plan:
Started OpendTect, selecting only PRO and DipSteering. Close it.
Switched on the environment variable OD_SKIP_PLUGIN_SEL, started od_main, all commercial
plugins are loaded.

Reviewers: nanne

Reviewed By: nanne

Subscribers: raman

Differential Revision: https://phabricator.dgbes.com:8443/D14978