Setting the appropriate location of well name annotations
The markers and horizon displays on top of the synthetic/seismic
viewer in the well tie window are polygons with a size of 2 (to
span accross multiple traces).

Setting the required alignment for that kind of data:
Left, from the first pick for well markers (left set of traces),
Right, and before the last pick for horizons (right set of traces),
both below the marker for readability.

[Jira: OD-3999]