Fixed well marker names not shown in SynthRock profile model viewer
Summary:
FlatView::AuxData uses a uiGraphicsItemGroup,
for which the ZValue is set before items are added to it. As a consequence
the added items did not have the ZValue of the group, and some of them are
thus not always visible. Here I propose that all items added to the group get by
default the ZValue of the group. The opposite was already in place: Setting
the ZValue of the group sets it for all its children.

Also removed a if ( not ) else, and extended your previous fix for uiTextItem to
uiAdvancedTextItem, also for now we do not have any that ignores transformations.

[Jira: OD-3999]

Test Plan: marker names are again showed in the SynthRock profile model viewer.

Reviewers: nanne

Reviewed By: nanne

Differential Revision: https://phabricator.dgbes.com:8443/D14997