made Interpolation available even when grids are matching
Summary: In 6.6.7 we had disabled the interpolation options when the incoming file had the same grid (B2C transform) as the host survey. Now added a checkbox to make the option available. It can be used when the inl/crl steps of the cbvs file > 1.

Test Plan: Interpolation option is always checkable in GUI. It just turns off by default when the grids are matching.

Reviewers: nanne

Reviewed By: nanne

Differential Revision: https://phabricator.dgbes.com:8443/D15008