Support Depth Horizons in Time surveys (with ZDomain info in omf file)
Summary:
Importing a depth horizon in depth domain is now supported in Time surveys. This revision includes:

1) UI change, where user can specifiy that the depth hor should be imported in depth domain itself, without any transformation.

2) Z transformation is disabled to achieve this.

3) For this to work, horizon should be aware of its zdomain, hence EMObject now has a boolean for this.

TO-DO: Reading back such horizons.

Test Plan: Tested the values by storing in ASCII format, and the z-unit specified in .hor file, in surfaces. More tests may be required, and possibly a test program as well.

Reviewers: raman

Reviewed By: raman

Subscribers: lloyd, wayne

Differential Revision: https://phabricator.dgbes.com:8443/D14975