added a notifier [devel]
Summary:
The objectInserter call was being made which lead to full
update of the object selection group. This caused crash in machine
learning becasue it was trying to access the ioobj which was not present
in machine learning list. Updating this list from callbacker function
did not work well in windows and did not allow added object to be
selected, hence this solution OD-4005

Test Plan: Now the object is added and can be selected on windows

Reviewers: arnaud

Reviewed By: arnaud

Subscribers: raman, lloyd, wayne

Differential Revision: https://phabricator.dgbes.com:8443/D15035