Correct Batch Process Name being used [devel]
Summary:
Earlier the setup name was being used. This led to creation of
file with that name. This lead to conflict because on windows multiple
process can not access writing to single file. On linux it worked fine

Test Plan:
Now able to launch multiple batch process on windows.
When forced to access the same log file twice, the log file in use error message was displayed.

Reviewers: arnaud

Reviewed By: arnaud

Subscribers: raman, lloyd, wayne

Differential Revision: https://phabricator.dgbes.com:8443/D15006