Marker filter should know the Depth type the User is using.
Summary: While marker filter in the Well table toolbar uses Depth type as one of the criteria, the same filter in the filter dialog didn't factor in the depth type. This revision introduces that as a parameter. DepthType cannot be chosen by the user in Well table filter, rather is set to whatever is chosen in the table window itself

Test Plan: tested in well table

Reviewers: arnaud

Reviewed By: arnaud

Subscribers: lloyd, wayne

Differential Revision: https://phabricator.dgbes.com:8443/D15066