Use reference counting for DataPacks
Summary:
Finally got odmain to compile after porting the od6.7 DataPack implementation.
Haven't modified any of the dgb plugins at the moment so a lot of these that use
DataPacks don't load.
A lot of testing and bug fixing is needed. Maybe you can check the DataPack API is what you want.

Using a weak pointer set in the DataPackMgrs to store the DataPack references is very agressive.
As soon as the reference count drops to zero the DataPack will be deleted. The applications use of
DataPacks needs to be adjusted so this doesn't happen. The issue is that the present use of DataPacks
assumes they will live on in the managers until explicitly deleted even if the DataPack goes out
of scope in the application. At the moment data put into the managers is disappearing before it
can be used because a reference to the DataPack isn't kept in an outer scope.

Test Plan: Compiles but things generally aren't working

Reviewers: arnaud

Reviewed By: arnaud

Differential Revision: https://phabricator.dgbes.com:8443/D15079