Changed return type for FlatView::Viewer::getPack [devel]
Returning a WeakPtr should make it clearer that the called is
not supposed to keep the datapack more than within the scope of
the current function.

The former return type can easily be retrieved by adding '.get()',
while the returned object can already be used to test if it contains
a pointer without having to explicitely ref and unref the object.

Tested display of lines in the scene, pre-loading of cubes.