Moved CRS/uiCRS to plugins [devel]
Summary:
Half of the applications we build, and all test programs but one do not need
the dependency to either proj or sqlite. Hence the CRS system should better stay a plugin
like it was before.
With few additional functions all existing code in uiTools is working, keeping the
position conversion dialog, which had been unified, in that module. Since we no longer
create another CRS position dialog, uiCRS is no longer dependend on uiODMain, but only on
uiTools.

reverts https://phabricator.dgbes.com:8443/D13397

Test Plan:
configured the solution with an without proj. No change seen with proj set on,
but the position conversion dialog behaves better than before this change when proj is off
(the proj dependent items are not added to the GUI).

Reviewers: nanne

Reviewed By: nanne

Subscribers: raman

Differential Revision: https://phabricator.dgbes.com:8443/D15131