Initial commit of the ReflCalc1D factory [devel]
Summary:
First commit of the RayTracer like factory, but limited to
AI, EI reflection coefficients.
Contains a RayTracerRunner equivalent as well, to easily compute them
from a set of ElasticModels
Eventually AICalc1D will be used for existing the Zero-offset synthetic computer,
as opposed to the VrmsRayTracer1D with a single offset.

Test Plan:
existing test programs are expanded, and return the same values
as computed in a test spreadsheet.

Reviewers: nanne

Reviewed By: nanne

Differential Revision: https://phabricator.dgbes.com:8443/D15138