Made RandomLine geometry based on Coord instead of BinID
Summary:
Next step is to make the required changes to randomline display classes.
This is an old request from Mike Pelissier

Test Plan:
Just tested adding a random line to the 3D scene and manipulate it a bit.
Goal is being able to extract a good looking randomline along a deviated well track.

Reviewers: raman

Reviewed By: raman

Differential Revision: https://phabricator.dgbes.com:8443/D15178