Added tool in SetBinding to set the application name [devel]
A fourth argument can be set now to provide the full path to the
binding library that needs to initialize the OpendTect application.
Since the application is setup from an external library and not from
an application, the shortest basename of that filepath is used to
set the application name, returned by any call to GetExecutableName().