Added tools for faciliate the integration of odpybind [devel]
Added tool in SetBinding to set the application name:
A fourth argument can be set now to provide the full path to the
binding library that needs to initialize the OpendTect application.
Since the application is setup from an external library and not from
an application, the shortest basename of that filepath is used to
set the application name, returned by any call to GetExecutableName().

Added odpybind to the auto-load list of the CRS plugin.
Added odpybind to the package definitions:
Using a global expression to collect all installed odpybind bindings, in case
several of them have been build against different versions of python.

Added path to the OpendTect dlls when setting the bindings:

CMake+Qt take care of adding the path to the DLLs for all OpendTect applications,
however that is not done for external bindings.
Using the windows API to add the OpendTect libraries directory to the search path,
without having to alter the PATH environment variable.

Now the odpybind module can successfully load all required DLLs, including the ones
from dependent plugins (CRS).