3D interation, FaultStickSets -> Faults [devel]
Summary:
Some loops were based on section IDs and the return value was
not being caught properly. A check has been placed to model the correct
behviour.
This occured when faultstickset displayed in 3D scene were converted to
faults

Test Plan:
Tested, the selected sticks are not present for processing.
Fault surface is still not being created

Reviewers: arnaud

Reviewed By: arnaud

Subscribers: raman, lloyd, wayne

Differential Revision: https://phabricator.dgbes.com:8443/D15213