Printing the helpKey in hexidecimal format when needed
SUMMARY
In debug mode the keys are displayed in format [provider] - [integer],
however this is now convenient for matching the content of the header files,
where the integer is displayed in hexadecimal format.
Hence added a conversion of that integer into its hexadecimal string.

TEST PLAN
hexadecimal string is now displayed, exactly like in include/General/od_helpids.h
Not too sure on how it will behave with the video helpIDs, as they do not show up in debug mode
(neither before nor after this patch).