Fix SEGFAULT when running odpybind tests [devel]
Summary:
Found the cause of the SEGFAULT when running the odpybind tests. The proposed change stops
storage of a CoordSystem reference in the SEGY translator. Issues arise when odpybind changes the
DBMDataSource which invalidates the CoordSystem reference set during loading modules. Attempting to
delete this invalid refptr at the end of the tests caused the crash.

Test Plan: Tested on Linux: odpybind tests run without SEGFAULT

Reviewers: arnaud, raman

Reviewed By: arnaud

Differential Revision: https://phabricator.dgbes.com:8443/D15225