Added option to unload a plugin [devel]
SUMMARY

Nice to have, but also fixed the implementation of PluginManager:load,
as a third case was not taken into account and thus not implemented:
It corresponds to the case of an auto-loaded plugin, which fails to load (for any reason):
findDataWithDispName will return nullptr for such plugin, even though it is present in the
data_ set.
The lack of implementation of that case made it such that the new PluginManager::Data object
created at the start of PluginManager:load was added to data_, instead of editing the
existing PluginManager::Data object for that plugin.

TEST PLAN

tested from closed source plugins. The data_ set is now edited in place, instead of
adding a new item for the existing plugin.