Fix for missing data for some 2D points in NN table
Summary: This is a temporary fix for 6.6. The immediate problem was the wrong assumption that the DataPointSet is sorted in GeomId/TrcNr. I have kept the code for synthetic data unchanged.

Test Plan: Created 2D NN with the same PointSets that gave a unch of empty rows earlier. Now all the rows are filled with valid data values.

Reviewers: nanne

Reviewed By: nanne

Differential Revision: https://phabricator.dgbes.com:8443/D15239