Fix for download errors in the Installer
Summary: The FileDownloader and the HttpRequestProcess operate in separate threads. There are situations where HttpRequestProcess gets data after the last read call made by FileDownloader and before setting the Finished status. That data is lying there in HttpRequestProcess::receiveddata_ and was getting lost as it was missed by the FileDownloader, leading to incomplete files being downloaded by the Installer.

Test Plan: Tested the Installer in the debug environment multiple times. No more failures.

Reviewers: arnaud

Reviewed By: arnaud

Differential Revision: https://phabricator.dgbes.com:8443/D15248