Added return to uiODMenuMgr::add2D3DMenuItem
Summary:
That function was void, leaving no easy way to retrieve which
action ID(s) had been effectively added. In that case where two items
are added using a submenu, it is only possible to return the action ID
of one of the two items.
This should be sufficient to locate the other if needed, since they are
added to the same submenu.
Looking at uiAction::getMenu(), I adjusted the implementation since it would
return nullptr for such 2D/3D sub menu, which I think is not correct because
there is an actual uiMenu being added and the parentactioncontainer_ is
correctly set.

Test Plan: compiles, to be used by plugins.

Reviewers: nanne, wayne

Reviewed By: wayne

Subscribers: raman

Differential Revision: https://phabricator.dgbes.com:8443/D15262