Fixed launching of terminal window on windows
Summary:
for a system installation of python the 'python enabled'
terminal did not start prior to the first bug fix for OD-4170.
Now that this part has been resolved, the terminal did start but in
background. Adding the flag '--uiconsole' to the command line for
the two terminals that needs it, cmd.exe and powershell.exe.
This flag is now implemented as well in the isolation script.
I am not using for this purpose OS::CommandExecPars, as its
implementation is less good than in odExternal library.

Test Plan: see child revision

Reviewers: raman, wayne

Reviewed By: wayne

Subscribers: nanne

Differential Revision: https://phabricator.dgbes.com:8443/D15272