Locks should never be copied
Summary:
Background: This change was triggered by a crash while creating a survey from a SEGY file. At some point a SEGY::FilePars object was being copied from another. Since SEGY::FilePars was recently turned into a CallBacker, the CallBacker::attachednotifierslock_ object was also copied. As a result they shared the same QRecursiveMutex pointer/object leading to a crash while destructing the second object.

IMHO locks are more of a part of the container than a part of the contents. Therefore when you copy the contents of an object in an assignment operator implementation, the locks should never be copied. Therefore I am flagging the copy constructors and assignment operators as deleted for all our Lock classes.

Another point, any class which is complex enough to contain a pointer (or ObjectSet for that matter) should have its own assignment operator implementation, as the default one will surely do the wrong thing. We probably need to look for more such classes and implement their assignment operators, but for now CallBacker surely is a candidate.

Test Plan: No more crash during survey setup. No impact on other functionalities observed so far.

Reviewers: nanne, arnaud

Reviewed By: nanne

Differential Revision: https://phabricator.dgbes.com:8443/D15277