Fixed some DataPack ref/unRef issues
Summary: In some instances, a RefMan was set to nullptr (triggering an unRef) and on top that there was an explit unRef, leading to crashes. Just went through the latest changes to fix most of those. Also tried to correct some other ref/unRef issues that were visible.

Test Plan: It's hard to test each of these, but I did not find any crashes after this, when I tried some randome workflows and 2D/3D displays.

Reviewers: wayne

Reviewed By: wayne

Subscribers: nanne

Differential Revision: https://phabricator.dgbes.com:8443/D15289