Displaying correct crl/trcnr in browse table [devel]
Summary:
The trace number was being assigned to crl number in case of 3D
Cube and that was giving wrong display of column header. Same was case
with 2d Seis browser. The display is now correct after the fix

Test Plan:
Display of column header, goto position window and wiggle
display is correct

Reviewers: arnaud

Reviewed By: arnaud

Subscribers: raman, lloyd, wayne

Differential Revision: https://phabricator.dgbes.com:8443/D15330