Added protections to IOMan
Summary:
several places within IOMan and Wavelet assumed there was always an
active, valid survey set in SI, especially IOMan::to should not try to look
for a survey subfolder without survey in the first place.

Added a mapping TrcKey <-> Coord for synthetic traces in the rare case it is
used without an active survey, like when running the test program test_stratsynthdatamgr
on build machines for some configurations.
However I retain the current implementation, with some bug fixes, if an active
survey is present, to disambiguate coordinates from the 3D survey from synthetic
traces coordinates.
Made a corresponding test case in test_trckeyzsampling.

Test Plan:
Both test programs are successfull, with and without valid project (root)
in the current user settings. od_main can be started and it is possible to switch
between projects.

Reviewers: nanne

Reviewed By: nanne

Subscribers: raman

Differential Revision: https://phabricator.dgbes.com:8443/D15331