Fix OD4239 Calculating horizon data from 2D is not working
Summary:
Fixes OD-4239 Calculating horizon data from 2D is not working
This is the batch process started from the "Processing|Create Horizon Output|Attribute|2D" menu
item. It doesn't appear to have been working even in 6.6.8. On Linux it crashes for me though Hesham
doesn't see this on Windows. Not sure if the override of the virtual useCoords() function will
break the ABI.

Test Plan: Tested on Linux

Reviewers: arnaud

Reviewed By: arnaud

Differential Revision: https://phabricator.dgbes.com:8443/D15363