Browse/Edit cube changes don't save
Summary:
The browse edit of the cube was not been saved and process was
going into infinite loop.
OD-4272

Test Plan:
Not hanging, saving the data [saved data is corrupted in cases when we try to edit first bin id position, if any position in between is chosen it is storing the data]. But the traces which are edited are removed
Upon saving the time values are corrupted. Earlier the time interval was 0-4-8.... but after saving it becomes 0-1000-2000.....

Reviewers: arnaud

Reviewed By: arnaud

Subscribers: raman, lloyd, wayne

Differential Revision: https://phabricator.dgbes.com:8443/D15388