Fixed listening servers not being always available [devel]
ServiceMgrBase::theMain is supposed to keep and provide a unique listening server,
to ensure that a single listening server is used by application.
However there are two types of listening servers: TCP/IP and 'Local'. Hence this
function should keep not one but two pointers, one for each possible type.

Fixes issues in closed-source plugins, because the BatchServiceClientMgr
global object was not OK: This service is using the 'local' listening server,
but the first one being initialized was a TCP/IP server.