Fixed initialization of uiIOObjSelGrp [devel]
Summary:
The filling of the uiListBox with uiIOObjSelGrp::fullUpdate should not be done
in the constructor, but on post-finalize, when all plugins are loaded and all base modules
are initialized.
Fixed also the default dataset not being set correctly the some translator groups, like the
seismic translator: The function TranslatorGroup::getSurveyDefaultKey must be used,
otherwise IOObj::isSurveyDefault will take the first available dataset, which can be
a SteeringCube.
It is unlikely that a SteeringCube will be a good default for a selection that lists both
'normal' seismic datasets and tagged datasets like SteeringCubes or FL/TFL.

Test Plan:
tested from the ML plugin, where the '4 Dip-steered median filter' dataset is now
the current item when the GUI is first displayed.

Reviewers: nanne

Reviewed By: nanne

Subscribers: raman

Differential Revision: https://phabricator.dgbes.com:8443/D15416