Seismic Brower UI update
Summary:
Solves OD-4272
Now the default mode for opening browser is readonly. If the cube is set to be locked, the option to edit the data will not be provided.
When the edit is saved, the save/saveas button is toggled off. This will be toggled on when the table is modified again.
The edit button will be disabled when edit mode is disabled.
Dialog type is set to close only

Test Plan: Tested by creating new volume and updating the same volume

Reviewers: arnaud

Reviewed By: arnaud

Subscribers: raman, lloyd, wayne

Differential Revision: https://phabricator.dgbes.com:8443/D15428